Rename ducks/tags -> ducks/allTags to reduce confusion with table tags
