+++ amundsen_application/static/js/ducks/tableMetadata/owners/sagas.ts
@@ -1,13 +1,9 @@
 import { call, select, takeEvery } from 'redux-saga/effects';
 import { SagaIterator } from 'redux-saga';
 
-import {
-  UpdateTableOwner, UpdateTableOwnerRequest,
-} from './reducer';
+import { UpdateTableOwner, UpdateTableOwnerRequest } from './reducer';
 
-import {
-  metadataUpdateTableOwner,
-} from '../api/v0';
+import { metadataUpdateTableOwner } from '../api/v0';
 
 // updateTableOwner
 export function* updateTableOwnerWorker(action: UpdateTableOwnerRequest): SagaIterator {
